+++ pr0gramm/__init__.py
@@ -1,5 +1,5 @@
 __version__ = '0.2.6'
 __author__ = "itssme https://github.com/itssme"
 
-
+from .items import *
 from .api import *
